Update README.md with a note about event bug in tx
We noticed that under certain conditions Z can send an event twice. This doesn't seem to be a problem anywhere in Webaverse except where we were having the avatar trigger twice, and rearranging the code fixed that. Adding a note in case someone else comes across this issue in the future. I don't think that we necessarily should change ZJS by adding a check unless it becomes a problem since it works as expected in the app if you follow the convention. I wouldn't even necessarily describe it as a bug.